+++ CrownsGuard.Multiplayer/Utilities/UnlockedFiguresHelper.cs
@@ -8,6 +8,6 @@ public static class UnlockedFiguresHelper
     public static bool IsValid(this BoardBlueprint boardBlueprint, byte[] unlockedFigures)
     {
         var bitArray = new BitArray(unlockedFigures);
-        return boardBlueprint.Figures.All(x => bitArray[x.FigureType]);
+        return boardBlueprint.Figures.All(x => bitArray[(int)x.FigureType]);
     }
 }
